[ticket] fix: 빌드 이슈 해결
